app/requirements.txt
streamlit==1.36.0
tensorflow==2.15.0
pillow
numpy
matplotlib
scikit-learn
protobuf==3.20.3
h5py
